requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
PyYAML==3.10
Werkzeug==0.9.3
appdotnet==0.0.1
itsdangerous==0.23
oauthlib==0.5.1
requests==1.2.3
requests-oauthlib==0.3.2
twython==3.0.0
wsgiref==0.1.2
